# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Zord backend server.",
        "title": "Swagger Zord API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/users": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "List Users",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "valid int",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "JWT Auth",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services_users_LIST.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services.Error"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Create Users",
                "parameters": [
                    {
                        "description": "body model",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services_users_CREATE.Data"
                        }
                    },
                    {
                        "type": "string",
                        "description": "JWT Auth",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services_users_CREATE.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services.Error"
                        }
                    }
                }
            }
        },
        "/users/{user_id}": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Delete Users",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Users ID",
                        "name": "users_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "JWT Auth",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services_users_DELETE.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services.Error"
                        }
                    }
                }
            }
        },
        "/users/{users_id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get a Users",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Users ID",
                        "name": "users_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "JWT Auth",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services_users_GET.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services.Error"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Edit Users",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Users ID",
                        "name": "users_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "body model",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services_users_EDIT.Data"
                        }
                    },
                    {
                        "type": "string",
                        "description": "JWT Auth",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services_users_EDIT.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "go-skeleton_internal_application_services.Error": {
            "type": "object",
            "properties": {
                "Error": {
                    "type": "string"
                },
                "Message": {},
                "Status": {
                    "type": "integer"
                }
            }
        },
...
        "go-skeleton_internal_application_services_users_CREATE.Data": {
            "type": "object",
            "required": [
                "Email",
                "Name",
                "Password",
                "Roles"
            ],
            "properties": {
                "Email": {
                    "type": "string"
                },
                "Id": {
                    "type": "string"
                },
                "Name": {
                    "type": "string"
                },
                "Password": {
                    "type": "string"
                },
                "Roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "go-skeleton_internal_application_services_users_CREATE.Response": {
            "type": "object",
            "properties": {
                "Created": {
                    "type": "boolean"
                }
            }
        },
        "go-skeleton_internal_application_services_users_DELETE.Response": {
            "type": "object",
            "properties": {
                "Message": {
                    "type": "string"
                }
            }
        },
        "go-skeleton_internal_application_services_users_EDIT.Data": {
            "type": "object",
            "properties": {
                "Email": {
                    "type": "string"
                },
                "Name": {
                    "type": "string"
                }
            }
        },
        "go-skeleton_internal_application_services_users_EDIT.Response": {
            "type": "object",
            "properties": {
                "Data": {
                    "$ref": "#/definitions/go-skeleton_internal_application_services_users_EDIT.Data"
                }
            }
        },
        "go-skeleton_internal_application_services_users_GET.Response": {
            "type": "object",
            "properties": {
                "Data": {
                    "$ref": "#/definitions/go-skeleton_internal_application_services_users_GET.UserResponse"
                }
            }
        },
        "go-skeleton_internal_application_services_users_GET.UserResponse": {
            "type": "object",
            "properties": {
                "Email": {
                    "type": "string"
                },
                "Id": {
                    "type": "string"
                },
                "Name": {
                    "type": "string"
                },
                "Roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "go-skeleton_internal_application_services_users_LIST.DataResponse": {
            "type": "object",
            "properties": {
                "Email": {
                    "type": "string"
                },
                "Id": {
                    "type": "string"
                },
                "Name": {
                    "type": "string"
                },
                "Roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "go-skeleton_internal_application_services_users_LIST.Response": {
            "type": "object",
            "properties": {
                "CurrentPage": {
                    "type": "integer"
                },
                "Data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/go-skeleton_internal_application_services_users_LIST.DataResponse"
                    }
                },
                "TotalPages": {
                    "type": "integer"
                }
            }
        }
    }
}